requirements.txt
alembic==1.17.2
bottle==0.13.4
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
fpsample<1.0.0
fsspec==2025.9.0
greenlet==3.2.4
h5py==3.15.1
joblib==1.5.2
kiwisolver==1.4.9
laspy==2.6.1
laszip==0.2.3
lazrs==0.8.0
mako==1.3.10
matplotlib==3.10.7
numpy==1.26.4
optuna==4.6.0
optuna-dashboard==0.20.0
packaging==25.0
pandas==2.3.3
pybind11==3.0.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sqlalchemy==2.0.44
threadpoolctl==3.6.0
torchinfo==1.8.0
tqdm==4.67.1
tzdata==2025.2
